requirements.txt
# NBA Stats Chat App dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web app
streamlit>=1.28.0

# LLM
anthropic>=0.18.0

# Database (CockroachDB)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization
plotly>=5.18.0
